refactor: use polymorphism to initialize TrashCompactor
When you look at the puzzle you notice the following:

- The operation initialization is the same in both parts.
- The result function is the same in both parts.

The only thing that changes is how the numbers are initialized for every operation.
Therefore, this refactor moves the result and operation initialization logic to a
parent abstract class, TrashCompactor; which also defines an abstract method
which responsability is just to initialize the numbers for every operation.